refactor(cli): extract runtime gate evaluation
